moved clean_cache() to script

+++ libs/cache.py
@@ -51,13 +51,12 @@ def sanitize_key(key):
     return "".join(rand.choices(VALID_FILE_CHARACTERS,k=10)) + '.json'
 
 def clean_cache():
-    for file in os.listdir(os.path.join('.','cache')):
-        mark_delete = False
-        with open(os.path.join('.','cache',file)) as f:
-            cache = json.load(f)
-            if cache['expires'] < time() - 60 * 60 * 6:
-                mark_delete = True
-        if mark_delete:
-            os.remove(os.path.join('.','cache',file))
-
-clean_cache()
+    if os.path.exists(os.path.join('.','cache')):
+        for file in os.listdir(os.path.join('.','cache')):
+            mark_delete = False
+            with open(os.path.join('.','cache',file)) as f:
+                cache = json.load(f)
+                if cache['expires'] < time() - 60 * 60 * 6:
+                    mark_delete = True
+            if mark_delete:
+                os.remove(os.path.join('.','cache',file))
